#feat
- 기능별 파일 분리
- amountStore 전역변수 관리

feat/refactoring
