+++ packages/yejchoi-vanilla/src/amount/amountStore.ts
@@ -1,0 +1,15 @@
+export const amountStore = (() => {
+    let totalAmount = 0
+
+    return {
+        getAmount : () => totalAmount,
+
+        setAmount : (value : number) => {
+            totalAmount = value
+        },
+
+        resetAmount : () => {
+            totalAmount = 0
+        }
+    };
+})();
