+++ packages/yejchoi-vanilla/src/types/types.ts
@@ -1,0 +1,5 @@
+export interface Product {
+    name : string;
+    price : number;
+    id : string;
+}
